sale details service creation

+++ back_end/src/main/java/lk/ijse/gdse/springboot/back_end/service/CustomerService.java
@@ -10,5 +10,5 @@ public interface CustomerService {
     void saveCustomer(CustomerDTO customerDTO);
     void updateCustomer(CustomerDTO customerDTO);
     void deleteCustomer(String id);
-    boolean updateTotalPoint(int point);
+    boolean updateTotalPoint(double point);
 }
